+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Admin settings: show and change all app settings via occ
+- Admins can delete the stored codes of one or all users via occ
 
 ## 3.2.0 (2026-07-05)